requirements.txt
streamlit
torch
transformers
diffusers
accelerate
Pillow
numpy
safetensors
